Move project_config.h to /include
